ejercicio 41 gestion de cursos acabado, aunque creo que he detectado un error en algun lado: Al utilizar el remove en los hashSet , cuando lo recorres con un for each y has de eliminar mas de uno, da un error. Imaginemos que tienes [0], [1] , [2] , y que quieres borrar el 1 y el 2, de alguna manera , cuando se borra el 1 , el 2 pasa a ser 1 , y cuando intenta acceder al 2 , da una excepcion

+++ _41_gestion_cursos/src/model/Curso.java
@@ -3,10 +3,10 @@ package model;
 public class Curso {
 	private String nombre;
 	private int duracion;
-	private int precio;
+	private double precio;
 	private String tematica;
 	
-	public Curso(String nombre, int duracion, int precio, String tematica) {
+	public Curso(String nombre, int duracion, double precio, String tematica) {
 		
 		this.nombre = nombre;
 		this.duracion = duracion;
@@ -30,11 +30,11 @@ public class Curso {
 		this.duracion = duracion;
 	}
 
-	public int getPrecio() {
+	public double getPrecio() {
 		return precio;
 	}
 
-	public void setPrecio(int precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
